Fix map pick and veto pagination on <=7 maps
Turns out that setting menu.Pagination = 7 won't work, since 7 is
apparently too high a number (i think there might be a miscalculation in
sourcemod on this). This can be seen by checking the return value
of calling SetMenuPagination (which will be false).

This is fixed by turning off pagination entirely if we have 7 maps or
less. I haven't added this logic to the side pick handler, as there
are never more than 2 options.

Offending code is either here:
https://github.com/alliedmodders/sourcemod/blob/237db0504c7a59e394828446af3e8ca3d53ef647/core/MenuStyle_Valve.cpp#L367
or here:
https://github.com/alliedmodders/sourcemod/blob/237db0504c7a59e394828446af3e8ca3d53ef647/core/MenuStyle_Radio.cpp#L598