Fixes veto's when using series_score for advantage
This is done by stopping the veto process once we hit the total series
score, and then adding maps to the front of the list of maps to play so
that they will be skipped.

There are a couple of downsides to this:
 - OnMapPicked and OnSidePicked map numbers are incorrect
 - It is no longer possible to use series_score + vetos as backup

I'm not 100% happy with this way of doing it now that i've tested it a
little, but it does fix our main problem.

Relates to: splewis/get5#119