Add configurable veto confirmation
Introduces a new setting get5_veto_confirmation_time, which is the time
in seconds from a menu has been presented to a choice is made, that will
result in a confirmation menu.

If set to 0, confirmations will never be shown. If set to a high number
(e.g. 3600 for an hour), confirmations will always be shown. Set to a
small value (e.g. 1 or 2) to prevent accidental vetos/picks, when a
player doesn't notice the menu while spamming buttons `1` through `5`.

Menu titles and options are customizable and translatable, and the
yes/no options are placed on button 8 and 9 respectively, to make sure
they won't get chosen by mistake.

Relates to: splewis/get5#122