Update natives.sp individual player name loading.
If the file exists, it will find the previous player name and replace the key. Otherwise a new file will be created to load all player names. I've tested with a few different cases of adding and removing various players from matches as well. Instead of reloading all the players, we only update the one. I think I was able to cover most tests in this case?