Adjusted the SetPlayer function to force name change. (#367)
* Adjusted the AddPlayer function.

It is now reloading the playernames in the given kV list, so users names can be custom once they are added, without having to reinit the match, or not have them load at all.

* Update logic on forcing team name to native.

* Updated to call LoadPlayers()

Cannot call just one player as it removes the previous players from the list.

* Update natives.sp individual player name loading.

If the file exists, it will find the previous player name and replace the key. Otherwise a new file will be created to load all player names. I've tested with a few different cases of adding and removing various players from matches as well. Instead of reloading all the players, we only update the one. I think I was able to cover most tests in this case?

* Format fixing

Copy paste sucks.

* Readjust back to LoadPlayerNames()