Add in timer for a fake client call as `ChangeClientTeam` breaks this functionality.
Add comments as to why we are using a timer with `FakeClientCommand`.

Per talks in sourcemod discord, use GetClientUserId and GetClientOfUserId to retrieve the client, this avoids any stacktraces being generated.

Remove debug log that was previously removed.

Remove unnecessary return.

Hook into Command_JoinTeam via timer when a user connects to the server, placing them on the correct team.

Update JoinTeam logic to check the buffer for any values as well.

Add in check to prevent unnecessary timer creation.

Reword comment.

Adjust JoinTeam to make sure we join the right team if we are not coaching.

Remove commented out code, remove TODO as its been tested fairly thoroughly.